Collectibles Prefabs and Knockback implemented

+++ Assets/Scripts/PlayerController/Stat Presets/PlayerBaseInfo.cs
@@ -7,8 +7,8 @@ public class PlayerBaseInfo : ScriptableObject
 [System.Serializable]
 public class PlayerData 
 {
-    public float CurrentHealth;
-    public float MaxHealth;
+    public int CurrentHealth;
+    public int MaxHealth;
 
     public ScriptableStats.Form Currentform;
